requirements.txt
astroid==2.1.0
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
cloudinary==1.15.0
coverage==4.5.2
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.7
django-braces==1.13.0
django-cors-middleware==1.3.1
django-extensions==2.1.6
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.9.1
gunicorn==19.9.0
requests==2.21.0
requests-oauthlib==1.2.0
idna==2.8
isort==4.3.9
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
more-itertools==6.0.0
oauthlib==3.0.1
pbr==5.1.3
pluggy==0.9.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.8.0
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
pylint==2.2.2
python-http-client==3.1.0
pytz==2018.9
sendgrid==5.6.0
pytest==4.3.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-core==3.1.0
urllib3==1.24.1
whitenoise==4.1.2
